raise when running some other unsafe ops during init
